Move ConvolutionProblem to Tensile client code (#713)
