+++ Tensile/Source/client/include/ReferenceValidator.hpp
@@ -32,7 +32,7 @@
 
 #include <Tensile/ContractionProblem.hpp>
 #include <Tensile/ContractionSolution.hpp>
-#include <Tensile/ConvolutionProblem.hpp>
+#include <ConvolutionProblem.hpp>
 
 #include "DataInitialization.hpp"
